GH-4 Moved call to addActions method
