+++ SwiftMonkey/MonkeyUIAutomation.swift
@@ -62,9 +62,9 @@ extension Monkey {
           of all relative probabilities.
     */
     public func addUIAutomationSingleTapAction(weight: Double) {
-        addAction(weight: weight, action: actInForeground { [weak self] in
+        addAction(weight: weight) { [weak self] in
             eventGenerator.sendTap(self!.randomPoint())
-        })
+        }
     }
 
     /**
@@ -85,7 +85,7 @@ extension Monkey {
     */
     public func addUIAutomationTapAction(weight: Double, multipleTapProbability: Double = 0.05,
     multipleTouchProbability: Double = 0.05, longPressProbability: Double = 0.05) {
-        addAction(weight: weight, action: actInForeground { [weak self] in
+        addAction(weight: weight) { [weak self] in
             let numberOfTaps: Int
             if self!.r.randomDouble()<multipleTapProbability {
                 numberOfTaps = Int(self!.r.randomUInt32() % 2) + 2
@@ -111,7 +111,7 @@ extension Monkey {
                 eventGenerator.liftUpAtPoints(touches, touchCount: UInt64(touches.count))
                 if i != numberOfTaps { self!.sleep(0.2) }
             }
-        })
+        }
     }
 
     /**
@@ -124,11 +124,11 @@ extension Monkey {
           of all relative probabilities.
     */
     public func addUIAutomationDragAction(weight: Double) {
-        addAction(weight: weight, action: actInForeground { [weak self] in
+        addAction(weight: weight) { [weak self] in
             let start = self!.randomPointAvoidingPanelAreas()
             let end = self!.randomPoint()
             eventGenerator.sendDragWithStartPoint(start, endPoint: end, duration: 0.5)
-        })
+        }
     }
 
     /**
@@ -141,11 +141,11 @@ extension Monkey {
           of all relative probabilities.
     */
     public func addUIAutomationFlickAction(weight: Double) {
-        addAction(weight: weight, action: actInForeground { [weak self] in
+        addAction(weight: weight) { [weak self] in
             let start = self!.randomPointAvoidingPanelAreas()
             let end = self!.randomPoint()
             eventGenerator.sendFlickWithStartPoint(start, endPoint: end, duration: 0.5)
-        })
+        }
     }
 
     /**
@@ -158,11 +158,11 @@ extension Monkey {
           of all relative probabilities.
     */
     public func addUIAutomationPinchCloseAction(weight: Double) {
-        addAction(weight: weight, action: actInForeground { [weak self] in
+        addAction(weight: weight) { [weak self] in
             let start = self!.randomPoint()
             let end = self!.randomPoint()
             eventGenerator.sendPinchCloseWithStartPoint(start, endPoint: end, duration: 0.5)
-        })
+        }
     }
 
     /**
@@ -175,11 +175,11 @@ extension Monkey {
           of all relative probabilities.
     */
     public func addUIAutomationPinchOpenAction(weight: Double) {
-        addAction(weight: weight, action: actInForeground { [weak self] in
+        addAction(weight: weight) { [weak self] in
             let start = self!.randomPoint()
             let end = self!.randomPoint()
             eventGenerator.sendPinchOpenWithStartPoint(start, endPoint: end, duration: 0.5)
-        })
+        }
     }
 
     /**
@@ -194,12 +194,12 @@ extension Monkey {
     */
     public func addUIAutomationRotateAction(weight: Double) {
         // Not working for some reason.
-        addAction(weight: weight, action: actInForeground { [weak self] in
+        addAction(weight: weight) { [weak self] in
             let point = self!.randomPoint()
             let radius = self!.r.randomDouble() * 100 + 50
             let angle = self!.r.randomDouble() * 2 * Double.pi
             eventGenerator.sendRotate(point, withRadius: radius, rotation: angle, duration: 0.5, touchCount: 2)
-        })
+        }
     }
 
     /**
@@ -212,7 +212,7 @@ extension Monkey {
           of all relative probabilities.
     */
     public func addUIAutomationOrientationAction(weight: Double) {
-        addAction(weight: weight, action: actInForeground { [weak self] in
+        addAction(weight: weight) { [weak self] in
             let orientations: [UIDeviceOrientation] = [
                 .portrait,
                 .portraitUpsideDown,
@@ -227,7 +227,7 @@ extension Monkey {
 
             eventGenerator.setOrientation(Int32(orientation.rawValue))
             self!.sleep(0.9)
-        })
+        }
     }
 
     /**
@@ -240,9 +240,9 @@ extension Monkey {
           of all relative probabilities.
     */
     public func addUIAutomationClickVolumeUpAction(weight: Double) {
-        addAction(weight: weight, action: actInForeground {
+        addAction(weight: weight) {
             eventGenerator.clickVolumeUp()
-        })
+        }
     }
 
     /**
@@ -255,9 +255,9 @@ extension Monkey {
           of all relative probabilities.
     */
     public func addUIAutomationClickVolumeDownAction(weight: Double) {
-        addAction(weight: weight, action: actInForeground {
+        addAction(weight: weight) {
             eventGenerator.clickVolumeDown()
-        })
+        }
     }
 
     /**
@@ -270,9 +270,9 @@ extension Monkey {
           of all relative probabilities.
     */
     public func addUIAutomationShakeAction(weight: Double) {
-        addAction(weight: weight, action: actInForeground {
+        addAction(weight: weight) {
             eventGenerator.shake()
-        })
+        }
     }
 
     /**
@@ -285,12 +285,12 @@ extension Monkey {
           of all relative probabilities.
     */
     public func addUIAutomationLockAction(weight: Double) {
-        addAction(weight: weight, action: actInForeground { [weak self] in
+        addAction(weight: weight) { [weak self] in
             let duration = 3 * self!.r.randomDouble()
             eventGenerator.lockDevice()
             self!.sleep(duration)
             eventGenerator.sendDragWithStartPoint(CGPoint(x: 20, y: 400), endPoint: CGPoint(x: 300, y: 400), duration: 0.5)
-        })
+        }
     }
 }
